feat: 관리자용 집계 서비스 구현 및 금액 데이터 Long으로 변경

+++ src/main/java/com/example/demo/domain/settlement/dto/SettlementSummaryItemDto.java
@@ -11,14 +11,14 @@ public class SettlementSummaryItemDto {
 	private Long creatorId;
 	private String creatorName;
 	private String yearMonth;
-	private Integer expectedSettlementAmount; // 예상 정산금액
+	private Long expectedSettlementAmount; // 예상 정산금액
 	private SettlementStatus status;
 
 	public static SettlementSummaryItemDto of(
 		Long creatorId,
 		String creatorName,
 		String yearMonth,
-		Integer expectedSettlementAmount,
+		Long expectedSettlementAmount,
 		SettlementStatus status
 	) {
 		return SettlementSummaryItemDto.builder()
