+++ src/main/java/com/example/demo/domain/settlement/dto/SettlementSummaryResponseDto.java
@@ -11,13 +11,13 @@ public class SettlementSummaryResponseDto {
 	private String startYearMonth;
 	private String endYearMonth;
 	private List<SettlementSummaryItemDto> settlements; // 모든 정산 데이터
-	private Integer totalExpectedSettlementAmount; // 전체 크리에이터 정산액 합계
+	private Long totalExpectedSettlementAmount; // 전체 크리에이터 정산액 합계
 
 	public static SettlementSummaryResponseDto of(
 		String startYearMonth,
 		String endYearMonth,
 		List<SettlementSummaryItemDto> settlements,
-		Integer totalExpectedSettlementAmount
+		Long totalExpectedSettlementAmount
 	) {
 		return SettlementSummaryResponseDto.builder()
 			.startYearMonth(startYearMonth)
